Barrierefreiheit 1: Add missing alt attributes. Empty for images that are decorative or convey just the same meaning than existing text.
Note that the alt texts served from the backend should, again, be equivalent to the meaning conveyed by the image. So if the image is just a teaser replicating in a more or less abstract way what the teaser text is already saying, the alt attribute should be empty.

Reference: https://www.w3.org/TR/WCAG20-TECHS/H37.html