1187: Fix error message on Firefox 'event is not defined'
that happens when starting to type on the searchbar textbox.

The cause was relying on window.event, which is not standard and not
implemented by Firefox. See https://developer.mozilla.org/en-US/docs/Web/API/Window/event

It is actually not related to the lack of output when the backend crashes, but it does mislead the analysis of such error scenarios.